improvement(k8s): better process mgmt and logging for dev mode sync
We now poll for more information from the mutagen daemon, and we've
overall improved the process management around it.

This unblocks more development around this feature, such as running
a postSyncCommand (like with the old hot reloading mechanism) and also
makes the use of mutagen easier for other contexts like syncing build
contexts for in-cluster building.